remove rollback case, move tests to relevant fixtures
